+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `read_from{_path}` will no longer take a `bool` for reading properties, and will do it by default. To
   change this behavior, you must now use `Probe`.
 
+## Removed
+- **lofty_attr**: The `#[lofty(always_present)]` attribute has been removed, and is now inferred.
+
 ## [0.8.1] - 2022-09-09
 
 ### Added